pyproject.toml
[tool.poetry]
name = "aiogram-template"
version = "0.1.0"
description = ""
authors = ["cortelf"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.9.0"
pydantic = "^1.10.2"
SQLAlchemy = "2.0.0b4"
aiogram = "3.0.0b6"
aiosqlite = "^0.18.0"
rodi = "^1.1.3"
psycopg = "^3.1.7"
internationalization-py = "0.1.4"
blacksheep = "^1.2.8"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
